extract queries to own files
